+++ inst/doc/v06-styling.R
@@ -1,13 +1,15 @@
 ## ---- message=FALSE, warning=FALSE, eval=FALSE---------------------------
 #  pivotStyles$addStyle(styleName="ColumnHeader", list(
-#      font="0.75em arial",
+#      "font-family"="arial",
+#      "font-size"="0.75em",
 #      padding="2px",
 #      border="1px solid blue",
 #      "vertical-align"="middle",
 #      "text-align"="center",
 #      "font-weight"="bold",
 #      color="blue",
-#      "background-color"="#FFFFFF"
+#      "background-color"="#FFFFFF",
+#      "xl-wrap-text"="wrap"
 #    ))
 
 ## ---- message=FALSE, warning=FALSE---------------------------------------
@@ -54,8 +56,6 @@ blue1Colors <- list(
   totalColor = "rgb(0, 0, 0)",
   borderColor = "rgb(48, 84, 150)"
 )
-# define the theme
-theme <- getSimpleColoredTheme(parentPivot=pt, colors=blue1Colors, fontName="Verdana, Arial")
 # create the pivot table
 library(pivottabler)
 pt <- PivotTable$new()
@@ -63,6 +63,8 @@ pt$addData(bhmtrains)
 pt$addColumnDataGroups("TrainCategory")
 pt$addRowDataGroups("TOC")
 pt$defineCalculation(calculationName="TotalTrains", summariseExpression="n()")
+# define the theme
+theme <- getSimpleColoredTheme(parentPivot=pt, colors=blue1Colors, fontName="Verdana, Arial")
 pt$theme <- theme
 pt$renderPivot(styleNamePrefix="t3")
 
@@ -77,8 +79,6 @@ grayColors <- list(
   totalColor = "rgb(0, 0, 0)",
   borderColor = "rgb(64, 64, 64)"
 )
-# define the theme
-theme <- getSimpleColoredTheme(parentPivot=pt, colors=grayColors, fontName="Courier New, Courier")
 # create the pivot table
 library(pivottabler)
 pt <- PivotTable$new()
@@ -86,6 +86,8 @@ pt$addData(bhmtrains)
 pt$addColumnDataGroups("TrainCategory")
 pt$addRowDataGroups("TOC")
 pt$defineCalculation(calculationName="TotalTrains", summariseExpression="n()")
+# define the theme
+theme <- getSimpleColoredTheme(parentPivot=pt, colors=grayColors, fontName="Courier New, Courier")
 pt$theme <- theme
 pt$renderPivot(styleNamePrefix="t4")
 
@@ -100,8 +102,6 @@ orangeColors <- list(
   totalColor = "rgb(0, 0, 0)",
   borderColor = "rgb(198, 89, 17)"
 )
-# define the theme
-theme <- getSimpleColoredTheme(parentPivot=pt, colors=orangeColors, fontName="Garamond, arial")
 # create the pivot table
 library(pivottabler)
 pt <- PivotTable$new()
@@ -109,6 +109,8 @@ pt$addData(bhmtrains)
 pt$addColumnDataGroups("TrainCategory")
 pt$addRowDataGroups("TOC")
 pt$defineCalculation(calculationName="TotalTrains", summariseExpression="n()")
+# define the theme
+theme <- getSimpleColoredTheme(parentPivot=pt, colors=orangeColors, fontName="Garamond, arial")
 pt$theme <- theme
 pt$renderPivot(styleNamePrefix="t5")
 
@@ -123,8 +125,6 @@ greenColors <- list(
   totalColor="rgb(0, 0, 0)",
   borderColor = "rgb(84, 130, 53)"
 )
-# define the theme
-theme <- getSimpleColoredTheme(parentPivot=pt, colors=greenColors, fontName="Helvetica, arial")
 # create the pivot table
 library(pivottabler)
 pt <- PivotTable$new()
@@ -132,6 +132,8 @@ pt$addData(bhmtrains)
 pt$addColumnDataGroups("TrainCategory")
 pt$addRowDataGroups("TOC")
 pt$defineCalculation(calculationName="TotalTrains", summariseExpression="n()")
+# define the theme
+theme <- getSimpleColoredTheme(parentPivot=pt, colors=greenColors, fontName="Helvetica, arial")
 pt$theme <- theme
 pt$renderPivot(styleNamePrefix="t6")
 
@@ -146,8 +148,6 @@ yellowColors <- list(
   totalColor="rgb(0, 0, 0)",
   borderColor = "rgb(255, 192, 0)"
 )
-# define the theme
-theme <- getSimpleColoredTheme(parentPivot=pt, colors=yellowColors, fontName="Verdana")
 # create the pivot table
 library(pivottabler)
 pt <- PivotTable$new()
@@ -155,6 +155,8 @@ pt$addData(bhmtrains)
 pt$addColumnDataGroups("TrainCategory")
 pt$addRowDataGroups("TOC")
 pt$defineCalculation(calculationName="TotalTrains", summariseExpression="n()")
+# define the theme
+theme <- getSimpleColoredTheme(parentPivot=pt, colors=yellowColors, fontName="Verdana")
 pt$theme <- theme
 pt$renderPivot(styleNamePrefix="t7")
 
@@ -169,7 +171,8 @@ createCustomTheme <- function(parentPivot=NULL, themeName="myCustomTheme") {
     ))
   # column headings in pink
   pivotStyles$addStyle(styleName="ColumnHeader", list(
-      font="0.75em \"Courier New\", Courier, monospace",
+      "font-family"="\"Courier New\", Courier, monospace",
+      "font-size"="0.75em",
       "font-weight"="bold",
       padding="2px",
       "border"="2px solid #B28DFF",
@@ -177,11 +180,13 @@ createCustomTheme <- function(parentPivot=NULL, themeName="myCustomTheme") {
       "text-align"="center",
       "font-weight"="bold",
       color="#DB49AC",
-      "background-color"="#FFCCF9"
+      "background-color"="#FFCCF9",
+      "xl-wrap-text"="wrap"
     ))
   # row headings in blue
   pivotStyles$addStyle(styleName="RowHeader", list(
-      font="0.75em \"Courier New\", Courier, monospace",
+      "font-family"="\"Courier New\", Courier, monospace",
+      "font-size"="0.75em",
       "font-weight"="bold",
       padding="2px 8px 2px 2px",
       "border"="1px solid #B28DFF",
@@ -189,11 +194,13 @@ createCustomTheme <- function(parentPivot=NULL, themeName="myCustomTheme") {
       "text-align"="left",
       "font-weight"="bold",
       color="#438EC8",
-      "background-color"="#ACE7FF"
+      "background-color"="#ACE7FF",
+      "xl-wrap-text"="wrap"
     ))
   # cells in yellow
   pivotStyles$addStyle(styleName="Cell", list(
-      font="0.75em \"Courier New\", Courier, monospace",
+      "font-family"="\"Courier New\", Courier, monospace",
+      "font-size"="0.75em",
       padding="2px 2px 2px 8px",
       "border"="1px solid #B28DFF",
       "text-align"="right",
@@ -202,7 +209,8 @@ createCustomTheme <- function(parentPivot=NULL, themeName="myCustomTheme") {
     ))
   # totals in orange
   pivotStyles$addStyle(styleName="Total", list(
-      font="0.75em \"Courier New\", Courier, monospace",
+      "font-family"="\"Courier New\", Courier, monospace",
+      "font-size"="0.75em",
       "font-weight"="bold",
       padding="2px 2px 2px 8px",
       "border"="1px solid rgb(84, 130, 53)",
@@ -239,8 +247,6 @@ orangeColors <- list(
   totalColor = "rgb(0, 0, 0)",
   borderColor = "rgb(198, 89, 17)"
 )
-# define the theme
-theme <- getSimpleColoredTheme(parentPivot=pt, colors=orangeColors, fontName="Garamond, arial")
 # create the pivot table
 library(pivottabler)
 pt <- PivotTable$new()
@@ -248,13 +254,15 @@ pt$addData(bhmtrains)
 pt$addColumnDataGroups("TrainCategory")
 pt$addRowDataGroups("TOC")
 pt$defineCalculation(calculationName="TotalTrains", summariseExpression="n()")
-# apply the theme
+# define the theme
+theme <- getSimpleColoredTheme(parentPivot=pt, colors=orangeColors, fontName="Garamond, arial")
 pt$theme <- theme
 # evaluate the pivot to generate the cells
 pt$evaluatePivot()
 # apply an additional highlight to one cell (3rd row, 2nd column)
+cellHighlight <- pt$createInlineStyle(declarations=list("background-color"="#FFFF00"))
 cells <- pt$getCells(specifyCellsAsList=TRUE, cellCoordinates=list(c(3, 2)))
-cells[[1]]$style <- PivotStyle$new(pt, "cellHighlight", list("background-color"="#FFFF00"))
+cells[[1]]$style <- cellHighlight
 # draw the pivot table
-pt$renderPivot(styleNamePrefix="t5")
+pt$renderPivot(styleNamePrefix="t9")